2025-AI/requirements.txt
Flask
Flask-SQLAlchemy
Pillow
numpy
opencv-python-headless
requests
torch
transformers
ftfy
regex
scikit-learn
python-dotenv